version pinning fixes
