+++ domains/apicredits/middleware/python/pyproject.toml
@@ -3,7 +3,7 @@ name = "arkhai-apicredits-middleware"
 version = "0.1.0"
 license = "MIT"
 description = "Seller-side ASGI middleware gating a downstream app on prepaid API credits from the arkhai credits service."
-requires-python = ">=3.12"
+requires-python = ">=3.12,<3.14"
 dependencies = [
   "httpx>=0.27",
 ]
