+++ domains/apicredits/sample-app/pyproject.toml
@@ -3,7 +3,7 @@ name = "arkhai-apicredits-sample-app"
 version = "0.1.0"
 license = "MIT"
 description = "Minimal token-gated HTTP service for the API-credits e2e topology."
-requires-python = ">=3.12"
+requires-python = ">=3.12,<3.14"
 dependencies = [
   "fastapi>=0.115.0",
   "uvicorn[standard]>=0.30.0",
