+++ domains/apicredits/storefront/pyproject.toml
@@ -3,7 +3,7 @@ name = "arkhai-apicredits-storefront"
 version = "0.2.3"
 license = "MIT"
 description = "Seller-side storefront for the API-credits domain: quota-backed listings, sync negotiation with ownership/quota guards, issuance-backed settlement."
-requires-python = ">=3.12"
+requires-python = ">=3.12,<3.14"
 dependencies = [
     "arkhai-core>=0.2.0",
     "arkhai-apicredits-domain>=0.1.0",
